Fix flatpak paths

+++ core/tabs/applications-setup/bottles-setup.sh
@@ -3,14 +3,9 @@
 . ../common-script.sh
 
 installBottles() {
-    if ! command_exists flatpak; then
-    printf "%b\n" "${YELLOW}Installing Bottles...${RC}"
-        case "$PACKAGER" in
-            *)
-                . ./setup-flatpak.sh
-                flatpak install -y flathub com.usebottles.bottles
-                ;;
-        esac
+    if ! flatpak_app_exists "com.usebottles.bottles"; then
+        printf "%b\n" "${YELLOW}Installing Bottles...${RC}"
+        flatpak install -y flathub com.usebottles.bottles
     else
         printf "%b\n" "${GREEN}Bottles is already installed.${RC}"
     fi
@@ -18,4 +13,5 @@ installBottles() {
 
 checkEnv
 checkEscalationTool
+checkFlatpak
 installBottles
